+++ src/files/client/custom/modules/photon-address/src/views/admin/photon-address-settings.js
@@ -24,6 +24,10 @@ define(['views/settings/record/edit'], function (Dep) {
                         {name: 'photonAddressUrl'},
                         {name: 'photonAddressLang'},
                     ],
+                    [
+                        {name: 'photonAddressBiasZoom'},
+                        {name: 'photonAddressBiasScale'},
+                    ],
                 ],
             },
         ]